+++ bs4/components.py
@@ -1,11 +1,7 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-from django.forms.widgets import flatatt
 from django.utils.safestring import mark_safe
-from bootstrap3.utils import render_tag
 
 from .text import text_value
+from .utils import render_tag
 
 
 def render_icon(icon, title=''):
@@ -13,7 +9,7 @@ def render_icon(icon, title=''):
     Render a Bootstrap glyphicon icon
     """
     attrs = {
-        'class': 'glyphicon glyphicon-{icon}'.format(icon=icon),
+        'class': f'glyphicon glyphicon-{icon}',
     }
     if title:
         attrs['title'] = title