+++ bs4/forms.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
 
 from django.contrib.admin.widgets import AdminFileWidget
 from django.forms import (
@@ -96,7 +94,7 @@ def render_button(
     elif size:
         raise BootstrapError(
             'Parameter "size" should be "xs", "sm", "lg" or ' +
-            'empty ("{}" given).'.format(size))
+            f'empty ("{size}" given).')
     if button_type:
         if button_type == 'submit':
             if not any([c.startswith('btn-') for c in split_css_classes(classes)]):
@@ -104,7 +102,7 @@ def render_button(
         elif button_type not in ('reset', 'button', 'link'):
             raise BootstrapError(
                 'Parameter "button_type" should be "submit", "reset", ' +
-                '"button", "link" or empty  ("{}" given).'.format(button_type))
+                f'"button", "link" or empty  ("{button_type}" given).')
         attrs['type'] = button_type
     attrs['class'] = classes
     icon_content = render_icon(icon) if icon else ''
@@ -142,8 +140,7 @@ def render_field_and_label(
         label_class = add_css_class(label_class, 'control-label')
     html = field
     if field_class:
-        html = '<div class="{klass}">{html}</div>'.format(
-            klass=field_class, html=html)
+        html = f'<div class="{field_class}">{html}</div>'
     if label:
         html = render_label(
             label, label_for=label_for, label_class=label_class) + html
@@ -154,10 +151,7 @@ def render_form_group(content, css_class=FORM_GROUP_CLASS):
     """
     Render a Bootstrap form group
     """
-    return '<div class="{klass}">{content}</div>'.format(
-        klass=css_class,
-        content=content,
-    )
+    return f'<div class="{css_class}">{content}</div>'
 
 
 def is_widget_required_attribute(widget):