+++ bs4/models.py
@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
 
 # Empty models.py, required file for Django tests