+++ bs4/text.py
@@ -1,11 +1,4 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-
-try:
-    from django.utils.encoding import force_text
-except ImportError:
-    from django.utils.encoding import force_unicode as force_text
+from django.utils.encoding import force_str
 
 
 def text_value(value):
@@ -14,7 +7,7 @@ def text_value(value):
     """
     if value is None:
         return ''
-    return force_text(value)
+    return force_str(value)
 
 
 def text_concat(*args, **kwargs):